Make BaseUserManage Page workable | Add SQLite Backup functionality

+++ BlazorBase.Backup/Services/BackupWebsiteService.cs
@@ -1,8 +1,8 @@
 ﻿using BlazorBase.Backup.Controller;
-using BlazorBase.CRUD.Services;
 using BlazorBase.Files.Models;
 using BlazorBase.MessageHandling.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Localization;
 using Microsoft.JSInterop;
 using System.Data.SqlClient;
@@ -11,15 +11,14 @@ using static BlazorBase.Backup.Controller.BlazorBaseBackupFileController;
 
 namespace BlazorBase.Backup.Services;
 
-public class BackupWebsiteService(DbContext dbContext, IStringLocalizer<BackupWebsiteService> localizer, IJSRuntime jsRuntime, IMessageHandler messageHandler)
+public class BackupWebsiteService(IConfiguration configuration, DbContext dbContext, IStringLocalizer<BackupWebsiteService> localizer, IJSRuntime jsRuntime, IMessageHandler messageHandler)
 {
     #region Injects
-
+    protected readonly IConfiguration Configuration = configuration;
     protected readonly DbContext DbContext = dbContext;
     protected readonly IStringLocalizer<BackupWebsiteService> Localizer = localizer;
     protected readonly IJSRuntime JSRuntime = jsRuntime;
     protected readonly IMessageHandler MessageHandler = messageHandler;
-
     #endregion
 
     public virtual async Task CreateAndDownloadWebsiteBackupAsync()
@@ -36,10 +35,15 @@ public class BackupWebsiteService(DbContext dbContext, IStringLocalizer<BackupWe
 
         DeleteOldBrokenBackups();
 
+        var databaseProvider = Configuration["DatabaseProvider"];
         var archivePath = Path.Join(tempFileStorePath, $"{Guid.NewGuid()}.WebsiteBackup.zip");
         using (var archive = ZipFile.Open(archivePath, ZipArchiveMode.Create))
         {
-            await CreateDatabaseBackupAsync(archive, progressId);
+            if (databaseProvider != null && databaseProvider == "SQLite")
+                await CreateSQLiteDatabaseBackupAsync(archive, progressId);
+            else
+                await CreateMSSQLDatabaseBackupAsync(archive, progressId);
+
             CreateFileStoreBackup(archive, progressId);
         }
 
@@ -48,7 +52,7 @@ public class BackupWebsiteService(DbContext dbContext, IStringLocalizer<BackupWe
         MessageHandler.CloseLoadingProgressMessage(progressId);
     }
 
-    protected virtual async Task CreateDatabaseBackupAsync(ZipArchive zipArchive, ulong progressId)
+    protected virtual async Task CreateMSSQLDatabaseBackupAsync(ZipArchive zipArchive, ulong progressId)
     {
         MessageHandler.UpdateLoadingProgressMessage(
             id: progressId,
@@ -81,15 +85,59 @@ public class BackupWebsiteService(DbContext dbContext, IStringLocalizer<BackupWe
             await Task.Delay(2000);
 
             zipArchive.CreateEntryFromFile(backupPath, Path.GetFileName(backupPath), CompressionLevel.Optimal);
-
-            File.Delete(backupPath);
         }
-        catch (Exception)
+        finally
         {
             if (File.Exists(backupPath))
                 File.Delete(backupPath);
+        }
+    }
 
-            throw;
+    protected virtual Task CreateSQLiteDatabaseBackupAsync(ZipArchive zipArchive, ulong progressId)
+    {
+        MessageHandler.UpdateLoadingProgressMessage(
+            id: progressId,
+            message: Localizer["Create Website Backup"],
+            progressText: Localizer["Creating Database Backup "],
+            currentProgress: 10
+        );
+
+        var backupPath = Path.Join(BlazorBaseFileOptions.Instance.TempFileStorePath, $"{Guid.NewGuid()}_DatabaseBackup.db");
+        var backupShmPath = Path.Join(BlazorBaseFileOptions.Instance.TempFileStorePath, $"{Guid.NewGuid()}_DatabaseBackup.db-shm");
+        var backupWalPath = Path.Join(BlazorBaseFileOptions.Instance.TempFileStorePath, $"{Guid.NewGuid()}_DatabaseBackup.db-wal");
+        
+        try
+        {
+            var connectionString = DbContext.Database.GetDbConnection().ConnectionString;
+            var connectionStringParts = connectionString.Split(";");
+            var dataSource = connectionStringParts.FirstOrDefault(x => x.Contains("Data Source"))?.Split("=")[1];
+            if (dataSource == null || !File.Exists(dataSource))
+                throw new Exception($"Can not find sqlite database file, specified in the connection strings data source: {dataSource}");
+
+            File.Copy(dataSource, backupPath, true);
+            zipArchive.CreateEntryFromFile(backupPath, Path.GetFileName(backupPath), CompressionLevel.Optimal);
+
+            var dbShmFilePath = dataSource + "-shm";
+            if (File.Exists(dbShmFilePath))
+            {
+                File.Copy(dbShmFilePath, backupShmPath, true);
+                zipArchive.CreateEntryFromFile(backupShmPath, Path.GetFileName(backupShmPath), CompressionLevel.Optimal);
+            }
+
+            var dbWalFilePath = dataSource + "-wal";
+            if (File.Exists(dbWalFilePath))
+            {
+                File.Copy(dbWalFilePath, backupWalPath, true);
+                zipArchive.CreateEntryFromFile(backupWalPath, Path.GetFileName(backupWalPath), CompressionLevel.Optimal);
+            }
+
+            return Task.CompletedTask;
+        }
+        finally
+        {
+            if (File.Exists(backupPath)) File.Delete(backupPath);
+            if (File.Exists(backupShmPath)) File.Delete(backupShmPath);
+            if (File.Exists(backupWalPath)) File.Delete(backupWalPath);
         }
     }
